refactor background & fix messenger issues
