+++ clients/extension/public/manifest.json
@@ -40,7 +40,7 @@
     "64": "icon64.png",
     "128": "icon128.png"
   },
-  "permissions": ["storage", "scripting", "activeTab"],
+  "permissions": ["storage", "activeTab"],
   "web_accessible_resources": [
     {
       "matches": ["<all_urls>"],
