+++ clients/extension/src/content/index.ts
@@ -23,9 +23,9 @@ try {
 }
 const inpageMessenger = initializeMessenger({ connect: 'inpage' })
 
-// There is a 1ms delay for initializing both sides of messengers
+// There is a 10ms delay for initializing both sides of messengers
 setTimeout(async () => {
   inpageMessenger.send('setDefaultProvider', {
     vultisigDefaultProvider: await getPrioritizeWallet(),
   })
-}, 1)
+}, 10)
